db: change database sqlite to mysql

+++ db/migrate/20250112100633_create_relationships.rb
@@ -1,8 +1,8 @@
 class CreateRelationships < ActiveRecord::Migration[8.0]
   def change
     create_table :relationships do |t|
-      t.integer :follower_id, null: false
-      t.integer :followed_id, null: false
+      t.bigint :follower_id, null: false
+      t.bigint :followed_id, null: false
 
       t.timestamps
     end
